feat: 위원회 페이지에서 로그인 유효한지 검증 로직 추가
