+++ src/hooks/committee/apply/useApplyDetail.ts
@@ -2,13 +2,17 @@ import { useGetPageParams } from "@/hooks/common/useGetPageParams";
 import { useApplyDetailPagination } from "@/hooks/committee/apply/useApplyDetailPagination";
 import { useGetApplyDetailQuery } from "@/hooks/tanstack-query/committee/event/useGetApplyDetail";
 import { useGetApplyDetailPageEventId } from "../apply/useGetApplyDetailPageEventId";
+import { ApiResponseError } from "@/types/common/api";
+import { useCommitteeCertification } from "@/hooks/committee/sign-in/useCommitteeCertification";
 
 export const useApplyDetail = () => {
   const { page: currentPage } = useGetPageParams();
   const { eventId } = useGetApplyDetailPageEventId();
 
   const { pagesPerGroup, queryParams, setPage } = useApplyDetailPagination(currentPage, eventId);
-  const { data, isLoading } = useGetApplyDetailQuery(queryParams);
+  const { data, isLoading, isError, error } = useGetApplyDetailQuery(queryParams);
+
+  useCommitteeCertification(isError, error as ApiResponseError);
 
   const ApplyDetailList = data?.content || [];
   const totalElements = data?.totalElements || 0;
