+++ src/hooks/committee/approve-wait/useApproveWaitList.ts
@@ -2,12 +2,15 @@ import { useGetPageParams } from "@/hooks/common/useGetPageParams";
 import { useApproveWaitPagination } from "./useApproveWaitPagination";
 import { useGetApproveWaitListQuery } from "@/hooks/tanstack-query/committee/approve-wait/useGetApproveWaitListQuery";
 import { usePostApprove } from "@/hooks/tanstack-query/committee/approve-wait/usePostApprove";
+import { ApiResponseError } from "@/types/common/api";
+import { useCommitteeCertification } from "@/hooks/committee/sign-in/useCommitteeCertification";
 
 export const useApproveWaitList = () => {
   const { page: currentPage } = useGetPageParams();
   const { pagesPerGroup, queryParams, setPage } = useApproveWaitPagination(currentPage);
 
-  const { data, isLoading } = useGetApproveWaitListQuery(queryParams);
+  const { data, isLoading, isError, error } = useGetApproveWaitListQuery(queryParams);
+  useCommitteeCertification(isError, error as ApiResponseError);
 
   const { onApprove } = usePostApprove();
