+++ src/hooks/tanstack-query/committee/announcement/useGetAnnouncementListQuery.ts
@@ -1,4 +1,6 @@
 import { getAnnouncementList } from "@/apis/committee/announcement";
+import { AnnouncementList } from "@/apis/dtos/committee/announcement";
+import { ApiResponseError } from "@/types/common/api";
 import { useQuery } from "@tanstack/react-query";
 
 interface AnnouncementListQueryParams {
@@ -8,7 +10,7 @@ interface AnnouncementListQueryParams {
 }
 
 export const useGetAnnouncementListQuery = ({ page, size, direction }: AnnouncementListQueryParams) => {
-  return useQuery({
+  return useQuery<AnnouncementList, ApiResponseError>({
     queryKey: ["announcementList", page, size, direction],
     queryFn: () => getAnnouncementList(page, size, direction),
     enabled: page >= 0,
