+++ src/hooks/tanstack-query/committee/approve-wait/useGetApproveWaitListQuery.ts
@@ -1,4 +1,6 @@
 import { getApproveWaitList } from "@/apis/committee/approve-wait";
+import { ApproveWaitList } from "@/apis/dtos/committee/approve-wait";
+import { ApiResponseError } from "@/types/common/api";
 import { useQuery } from "@tanstack/react-query";
 
 interface ApproveWaitListQueryParams {
@@ -8,7 +10,7 @@ interface ApproveWaitListQueryParams {
 }
 
 export const useGetApproveWaitListQuery = ({ page, size, direction }: ApproveWaitListQueryParams) => {
-  return useQuery({
+  return useQuery<ApproveWaitList, ApiResponseError>({
     queryKey: ["approveWait", page, size, direction],
     queryFn: () => getApproveWaitList(page, size, direction),
     enabled: page >= 0,
