+++ src/hooks/tanstack-query/committee/event/useGetApplyDetail.ts
@@ -1,4 +1,6 @@
 import { getApplyDetail } from "@/apis/committee/apply";
+import { ApplyDetailList } from "@/apis/dtos/committee/apply";
+import { ApiResponseError } from "@/types/common/api";
 import { useQuery } from "@tanstack/react-query";
 
 interface ApplyDetailQueryParams {
@@ -9,7 +11,7 @@ interface ApplyDetailQueryParams {
 }
 
 export const useGetApplyDetailQuery = ({ eventId, page, size, direction = "asc" }: ApplyDetailQueryParams) => {
-  return useQuery({
+  return useQuery<ApplyDetailList, ApiResponseError>({
     queryKey: ["applyDetail", eventId, page, size, direction],
     queryFn: () => getApplyDetail(eventId, page, size, direction),
     enabled: page >= 0 && !!eventId,
