+++ src/hooks/tanstack-query/committee/event/useGetApplyList.ts
@@ -1,4 +1,6 @@
 import { getEventList } from "@/apis/committee/event";
+import { EventList } from "@/apis/dtos/committee/event";
+import { ApiResponseError } from "@/types/common/api";
 import { useQuery } from "@tanstack/react-query";
 
 interface ApplyListQueryParams {
@@ -8,7 +10,7 @@ interface ApplyListQueryParams {
 }
 
 export const useGetApplyListQuery = ({ page, size, direction }: ApplyListQueryParams) => {
-  return useQuery({
+  return useQuery<EventList, ApiResponseError>({
     queryKey: ["applyList", page, size, direction],
     queryFn: () => getEventList(page, size, direction),
     enabled: page >= 0,
