+++ src/hooks/tanstack-query/committee/event/useGetEventList.ts
@@ -1,4 +1,6 @@
 import { getEventList } from "@/apis/committee/event";
+import { EventList } from "@/apis/dtos/committee/event";
+import { ApiResponseError } from "@/types/common/api";
 import { useQuery } from "@tanstack/react-query";
 
 interface EventListQueryParams {
@@ -8,7 +10,7 @@ interface EventListQueryParams {
 }
 
 export const useGetEventListQuery = ({ page, size, direction }: EventListQueryParams) => {
-  return useQuery({
+  return useQuery<EventList, ApiResponseError>({
     queryKey: ["eventList", page, size, direction],
     queryFn: () => getEventList(page, size, direction),
     enabled: page >= 0,
