+++ src/hooks/tanstack-query/common/my-info/useGetMyInfo.ts
@@ -1,10 +1,9 @@
 import { getMyInfo } from "@/apis/common/my-info";
 import { MyInfo } from "@/apis/dtos/common/my-info";
-import { ApiResponseError } from "@/types/common/api";
 import { useQuery } from "@tanstack/react-query";
 
 export const useGetMyInfo = () => {
-  return useQuery<MyInfo, ApiResponseError>({
+  return useQuery<MyInfo>({
     queryKey: ["myInfo"],
     queryFn: () => getMyInfo(),
   });
